Rename basics into validity. Minor optimizations.
